implemented accuracy calculation for regression data points

+++ LIMS/Model/RegressionModels/RegressionDataPoint.cs
@@ -10,7 +10,7 @@ namespace LIMS.Model.RegressionModels
         // Instrument Response
         public double? InstrumentResponse { get; init; }
         public double? NominalConcentration { get; set; }
-        public decimal? Accuracy { get; set; }
+        public double? Accuracy { get; set; }
 
         public bool IsActive { get; set; } = true;
         public string SampleName { get; init; }
